+++ CHANGES.md
@@ -1,6 +1,13 @@
-## 0.2.4.1 (2020-??-??)
+## 0.2.5.0 (2020-06-15)
 
- * TBD.
+ * Bugfix: when reporting counter values to statsd, send only the
+   increments ([#23](https://github.com/tibbe/ekg-statsd/pull/23)).
+
+## 0.2.4.1 (2020-05-21)
+
+ * Sanitize metric names by replacing `:` with `_` to adhere to the StatsD
+   protocol ([#26](https://github.com/tibbe/ekg-statsd/pull/26)).
+ * Add support for GHC 8.10 in base bounds and CI.
 
 ## 0.2.4.0 (2018-08-01)